+++ moderatorFrontend/src/jsMain/kotlin/views/accessManagement/accessManagementDetails/accessManagementDetailsActionButtons/AccessManagementDetailsActionButtonsConfig.kt
@@ -1,0 +1,11 @@
+package views.accessManagement.accessManagementDetails.accessManagementDetailsActionButtons
+
+import react.RefObject
+import views.accessManagement.accessManagementDetails.AccessManagementDetailsConfig
+import webcore.MbDialogRef
+
+data class AccessManagementDetailsActionButtonsConfig(
+  val accessManagementDetailsType: AccessManagementDetailsConfig,
+  val dialogRef: RefObject<MbDialogRef>,
+  val createAccessControlOnClick: () -> Unit,
+)
