+++ moderatorFrontend/src/jsMain/kotlin/views/accessManagement/accessManagementExport/accessManagementExportTableRow/AccessManagementExportTableRowConfig.kt
@@ -1,0 +1,7 @@
+package views.accessManagement.accessManagementExport.accessManagementExportTableRow
+
+import com.studo.campusqr.common.payloads.AccessManagementExportData
+
+class AccessManagementExportTableRowConfig(
+  val permit: AccessManagementExportData.Permit,
+)
