+++ moderatorFrontend/src/jsMain/kotlin/views/accessManagement/accessManagementExport/listAccessManagementExport/ListAccessManagementExportConfig.kt
@@ -1,0 +1,5 @@
+package views.accessManagement.accessManagementExport.listAccessManagementExport
+
+class ListAccessManagementExportConfig(
+  val locationId: String?,
+)
