+++ moderatorFrontend/src/jsMain/kotlin/views/accessManagement/accessManagementOverview/AccessManagementListConfig.kt
@@ -1,0 +1,5 @@
+package views.accessManagement.accessManagementOverview
+
+class AccessManagementListConfig(
+  val locationId: String?,
+)
