+++ moderatorFrontend/src/jsMain/kotlin/views/report/Report.kt
@@ -19,6 +19,8 @@ import views.common.CenteredProgressFc
 import views.common.MbLinearProgressFc
 import views.common.spacer
 import webcore.*
+import webcore.datePicker.DatePickerConfig
+import webcore.datePicker.DatePickerFc
 import webcore.NavigationHandler.dialogRef
 import kotlin.js.Date
