+++ moderatorFrontend/src/jsMain/kotlin/webcore/timePicker/TimePickerConfig.kt
@@ -1,0 +1,18 @@
+package webcore.timePicker
+
+import mui.material.FormControlVariant
+import kotlin.js.Date
+
+class TimePickerConfig(
+  var time: Date,
+  var onChange: (date: Date) -> Unit,
+  var disabled: Boolean = false,
+  var error: Boolean = false,
+  var min: Date? = null,
+  var max: Date? = null,
+  var stepMinutes: Int = 1,
+  var fullWidth: Boolean = false,
+  var label: String? = null,
+  var helperText: String? = null,
+  var variant: FormControlVariant = FormControlVariant.outlined,
+)
